De-duplicate code related to mysql_stmt_free_result_*.
The function `Nonblocking.Res.free` was not exposed in the API and not
used anywhere else.  There was, however, a duplicate implementation in
`Make.Stmt.free_res`, which has now been simplified after moving the
shared to `Stmt.free_result` in consistency with the MySQL identifier
namespace.

This resolves warning 32 from the `dune-workspace` overrides.